Add files via upload
Handwritten digit recognition is a classic problem in the field of computer vision and machine learning. The task involves classifying images of handwritten digits (0-9) into their respective categories. This project uses the CIFAR-10 dataset and the PyTorch deep learning framework to build and train a convolutional neural network (CNN) model for digit classification.